Fixed Gender Class
Took 2 hours 46 minutes
